Rename page files to match tab names.
